src/github/python/requirements.txt
mysql-connector-python==2.0.4
requests==2.31.0
numpy==1.24.2